sistem-pa/requirements.txt
fpdf2==2.8.5
openpyxl==3.1.2
python-docx==1.1.0
Flask==3.1.2
flask-sock==0.7.0
Werkzeug==3.1.5
flask-smorest==0.44.0
Flask-WTF==1.2.2
email-validator==2.2.0
flask-swagger
flask-swagger-ui
python-dotenv
